modified format of tests
